Implement Queue and Cell classes
Initialize a Queue class that
would be used to hold all
the adjacencies(valid moves)
of a node.

Set an default parameter of
the Queue's constructor and
assign the instance property
items, to the default
parameter of constructor.

Initialize the isEmpty,
enqueue, dequeue and size
methods in the Queue class.

Initialize the Cell class.
Every node will be a Cell.
Set the constructor
parameters. Set the x and y
instance properties to
the parameters of the
constructor. Set the
predecessors instance
property to be null.