Generate valid moves.
Initialize a private method
that takes in `coordinates`
and dynamically generates
valid moves for the cell
in the coordinates.

Deconstruct the `coordinates`
into variables.
Initialize an empty
`validMovesCellsArray`array.

Check if the probable
moves of the passed
in coordinates are
valid moves
within the board range
(greater than 0 and
less than 8). If they
are, instantiate a new Cell
with the moves and
push them into the
validMovesCellsArray.

Return the
validMovesCellsArray array.