Get the shortest path
Initialize a knightMoves
method that takes in
`startCoordinates` and
`endCoordinates` and
returns the shortest
possible path from the
cell in the startCoordinates
to the cell in the
endCoordinates.

First check if the
coordinates are valid
and return accordingly.

If valid, create a new
Queue and enqueue the
cell that is in the
startCoordinates.

While the Queue is
not empty, dequeue the
first element (currentElement)
and get its valid moves.

For each cell in the
array of valid moves, if
a cell does not have
`predecessors`, set its
predecessors to an empty
array. This is to keep
the predecessors of the
cell in the start
coordinates as null.

If the currentElement has
predecessors, for each of
its predecessors, push
each of them into the
predecessors of (each)
the cell
in the array of valid
moves. This is because
the currentElement is the
predecessor of all the
cells in the array
of valid moves and
all the predecessors of
currentElement are also
technically their
predecessors (ancestors).

Push the currentElement
into the the list of
predecessors of the cell
and enqueue the cell.

Look for the end coordinates
in the items array of the
Queue and if it is found,
return its predecessors.

In main.js, log the result of
calling the knightMoves()
method on certain start and end
coordinates.