Render the path of predecessors
Initialize a private method
renderPath that renders the
path of the predecessors.

The method takes in a
`foundCell` argument and
transforms each predecessor
in the foundCell's
array(of strings) of predecessors
to the cell's coordinates
with array.map method.
It calls join()
on the resulting array to
join each predecessor's
coordinates with a
`->`.

The method returns this
path and the
found cell's coordinates.

Update knightMoves to
call the renderPath()
on the destination
cell found in the Queue's
items array. Use the
length of the cell's
predecessor's array to
denote the number of
moves.

This returns the
predecessors
and the number of
moves.